requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-image>=0.21.0
matplotlib>=3.7.0
pandas>=2.0.0

# 可选：用于MusicXML导出
# music21>=9.0.0